requirements.txt
# requirements.txt
azure-functions
azure-storage-queue
azure-storage-blob